# apps/api/public/swagger.yml
swagger: "2.0"
info:
  description: "GitHub Release Notification API that allows users to subscribe to email notifications about new releases of a chosen GitHub repository."
  version: "1.0.0"
  title: "GitHub Release Notification API"
basePath: "/api"
...
schemes:
  - "https"
  - "http"
paths:
  /subscribe:
    post:
      tags:
        - "subscription"
      summary: "Subscribe to release notifications"
      description: "Subscribe an email to receive notifications about new releases of a GitHub repository. The repository is validated via GitHub API."
      operationId: "subscribe"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Subscription request"
          required: true
          schema:
            $ref: "#/definitions/SubscribeRequest"
      responses:
        "200":
          description: "Subscription successful. Confirmation email sent."
        "400":
          description: "Invalid input (e.g., invalid repo format)"
        "404":
          description: "Repository not found on GitHub"
        "409":
          description: "Email already subscribed to this repository"
...
definitions:
  SubscribeRequest:
    type: "object"
    required:
      - "email"
      - "repo"
    properties:
      email:
        type: "string"
        description: "Email address to subscribe"
      repo:
        type: "string"
        description: "GitHub repository in owner/repo format (e.g., golang/go)"
...